+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+## 10.1.0 (08-09-2026)
+
+### Bugfix: 1
+- [#41116](https://parermine.regione.emilia-romagna.it/issues/41116) STRUMENTI URBANISTICI: mancanza icone modifica e cestino in interfaccia PING
+
+### Novità: 3
+- [#41085](https://parermine.regione.emilia-romagna.it/issues/41085) Modifica a PING per gestire il numero di oggetti da mandare in trasformazione
+- [#40936](https://parermine.regione.emilia-romagna.it/issues/40936) Introduzione della cartella STAGING per il deposito di materiale da elaborare con le trasformazioni
+- [#40708](https://parermine.regione.emilia-romagna.it/issues/40708) Introdurre il nuovo framework con la modifica per l'accesso tramite CIE/SPID
+
 ## 10.0.0 (13-07-2026)
 
 ### Novità: 1